--- Improves #8---
* Moved class registration to be on per downlink level.
