+++ swimai/client/connections.py
@@ -105,6 +105,7 @@ class WSConnection:
     async def open(self) -> None:
         if self.status == ConnectionStatus.CLOSED:
             self.status = ConnectionStatus.CONNECTING
+            # TODO Add exception handling
             self.websocket = await websockets.connect(self.host_uri)
             self.status = ConnectionStatus.IDLE
 
@@ -158,7 +159,9 @@ class WSConnection:
         if self.websocket is None or self.status == ConnectionStatus.CLOSED:
             await self.open()
 
-        await self.websocket.send(message)
+        if self.websocket:
+            # TODO Add exception handling
+            await self.websocket.send(message)
 
     async def wait_for_messages(self) -> None:
         """
@@ -245,6 +248,8 @@ class DownlinkManager:
         self.connection = connection
         self.status = DownlinkManagerStatus.CLOSED
         self.downlink_model = None
+        self.registered_classes = dict()
+        self.strict = True
         self.__downlink_views = dict()
 
     @property
@@ -273,8 +278,8 @@ class DownlinkManager:
 
         :param downlink_view:       - Downlink view with the information about the remote agent.
         """
-        self.downlink_model = await downlink_view.create_downlink_model()
-        self.downlink_model.downlink = self
+        self.downlink_model = await downlink_view.create_downlink_model(self)
+        # self.downlink_model.downlink_manager = self
         self.downlink_model.connection = self.connection
 
     async def add_view(self, downlink_view: 'ValueDownlinkView') -> None:
