+++ test/client/test_connections.py
@@ -852,7 +852,7 @@ class TestConnections(unittest.TestCase):
         self.assertEqual(downlink_view.client, actual.downlink_model.client)
         self.assertEqual(downlink_view.node_uri, actual.downlink_model.node_uri)
         self.assertEqual(downlink_view.host_uri, actual.downlink_model.host_uri)
-        self.assertEqual(actual, actual.downlink_model.downlink)
+        self.assertEqual(actual, actual.downlink_model.downlink_manager)
         self.assertEqual(actual.connection, actual.downlink_model.connection)
 
     @patch('swimai.client.connections.WSConnection.send_message', new_callable=MockAsyncFunction)
@@ -876,7 +876,7 @@ class TestConnections(unittest.TestCase):
         self.assertEqual(downlink_view.host_uri, actual.downlink_model.host_uri)
         self.assertEqual(DownlinkManagerStatus.OPEN, actual.status)
         self.assertEqual(downlink_view.node_uri, actual.downlink_model.node_uri)
-        self.assertEqual(actual, actual.downlink_model.downlink)
+        self.assertEqual(actual, actual.downlink_model.downlink_manager)
         self.assertEqual(downlink_view.lane_uri, actual.downlink_model.lane_uri)
         self.assertEqual(actual.connection, actual.downlink_model.connection)
         self.assertEqual(1, mock_schedule_task.call_count)
@@ -913,7 +913,7 @@ class TestConnections(unittest.TestCase):
         self.assertEqual(first_downlink_view.node_uri, actual.downlink_model.node_uri)
         self.assertIsInstance(actual.downlink_model, ValueDownlinkModel)
         self.assertEqual(first_downlink_view.lane_uri, actual.downlink_model.lane_uri)
-        self.assertEqual(actual, actual.downlink_model.downlink)
+        self.assertEqual(actual, actual.downlink_model.downlink_manager)
         self.assertEqual(actual.connection, actual.downlink_model.connection)
         self.assertEqual(1, mock_schedule_task.call_count)
         self.assertEqual(1, mock_send_message.call_count)
@@ -943,7 +943,7 @@ class TestConnections(unittest.TestCase):
         self.assertEqual(DownlinkManagerStatus.CLOSED, actual.status)
         self.assertIsInstance(actual.downlink_model, ValueDownlinkModel)
         self.assertEqual(downlink_view.node_uri, actual.downlink_model.node_uri)
-        self.assertEqual(actual, actual.downlink_model.downlink)
+        self.assertEqual(actual, actual.downlink_model.downlink_manager)
         self.assertEqual(downlink_view.lane_uri, actual.downlink_model.lane_uri)
         self.assertEqual(actual.connection, actual.downlink_model.connection)
         self.assertEqual(2, mock_schedule_task.call_count)
@@ -979,7 +979,7 @@ class TestConnections(unittest.TestCase):
         self.assertIsInstance(actual.downlink_model, ValueDownlinkModel)
         self.assertEqual(DownlinkManagerStatus.OPEN, actual.status)
         self.assertEqual(actual.connection, actual.downlink_model.connection)
-        self.assertEqual(actual, actual.downlink_model.downlink)
+        self.assertEqual(actual, actual.downlink_model.downlink_manager)
         self.assertEqual(1, mock_schedule_task.call_count)
         self.assertEqual(1, mock_send_message.call_count)
         self.assertTrue(first_downlink_view.initialised.is_set())
